+++ GRDB/Core/DatabaseValue.swift
@@ -87,7 +87,7 @@ public struct DatabaseValue {
             return value
         }
         guard isNull else {
-            fatalError("could not convert \(self) to \(Value.self).")
+            fatalError("could not convert database value \(self) to \(Value.self)")
         }
         return nil
     }
@@ -100,7 +100,7 @@ public struct DatabaseValue {
     /// - returns: A *Value*.
     public func value<Value: DatabaseValueConvertible>() -> Value {
         guard let value = Value.fromDatabaseValue(self) as Value? else {
-            fatalError("could not convert \(self) to \(Value.self).")
+            fatalError("could not convert database value \(self) to \(Value.self)")
         }
         return value
     }