+++ GRDB/Core/Row.swift
@@ -372,7 +372,7 @@ extension Row {
     
     private static func statementColumnConvertible<Value: StatementColumnConvertible>(atUncheckedIndex index: Int, in sqliteStatement: SQLiteStatement) -> Value {
         guard sqlite3_column_type(sqliteStatement, Int32(index)) != SQLITE_NULL else {
-            fatalError("could not convert NULL to \(Value.self).")
+            fatalError("could not convert database NULL value to \(Value.self)")
         }
         return Value.init(sqliteStatement: sqliteStatement, index: Int32(index))
     }