+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 Release Notes
 =============
 
+## 0.71.0
+
+Released June 5, 2016
+
+**Fixed**
+
+- Fix a crash that would sometimes happen when a FetchedRecordsController's callbacks avoid retain cycles by capturing unowned references.
+- Improved handling of numeric overflows. Fixes [#68](https://github.com/groue/GRDB.swift/issues/68).
+
+
+**New**
+
+- GRDB can now use a custom SQLite build ([documentation](https://github.com/groue/GRDB.swift/tree/master/SQLiteCustom)). Merged [#62](https://github.com/groue/GRDB.swift/pull/62) by [@swiftlyfalling](https://github.com/swiftlyfalling).
+- With a custom SQLite build, transaction observers can observe individual column values in the rows modified by a transaction ([documentation](https://github.com/groue/GRDB.swift#support-for-sqlite-pre-update-hooks)). Merged [#63](https://github.com/groue/GRDB.swift/pull/63) by [@swiftlyfalling](https://github.com/swiftlyfalling).
+- FetchedRecordsController can now fetch other values alongside the fetched records. This grants you the ability to fetch values that are consistent with the notified changes. ([documentation](https://github.com/groue/GRDB.swift#the-changes-notifications))
+
+
+**Breaking Changes**
+
+- iOS7 is no longer supported.
+
+
 ## 0.70.1
 
 Released May 30, 2016